# cmd/agent/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: agent-sandbox-agent
  version: 1.0.0
  description: RESTful API exposed by the sandbox agent sidecar process.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/projects:
    get:
      tags: [projects]
      summary: List projects
      operationId: listProjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [projects]
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: '#/components/parameters/projectId'
    get:
      tags: [projects]
      summary: Get project
      operationId: getProject
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [projects]
      summary: Update project
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [projects]
      summary: Delete project
      operationId: deleteProject
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'

  # ── Sessions ──────────────────────────────────────────────────────────────
...
  /api/v1/terminals:
    get:
      tags: [terminals]
      summary: List terminals
      operationId: listTerminals
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [terminals]
      summary: Create terminal
      operationId: createTerminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/terminals/{terminalId}:
    parameters:
      - $ref: '#/components/parameters/terminalId'
    get:
      tags: [terminals]
      summary: Get terminal info
      operationId: getTerminal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [terminals]
      summary: Close terminal
      operationId: deleteTerminal
      responses:
        '204':
          description: Closed
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/terminals/{terminalId}/resize:
    parameters:
      - $ref: '#/components/parameters/terminalId'
    post:
      tags: [terminals]
      summary: Resize terminal
      operationId: resizeTerminal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalResizeRequest'
      responses:
        '204':
          description: Resized
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/terminals/{terminalId}/ws:
    parameters:
      - $ref: '#/components/parameters/terminalId'
    get:
      tags: [terminals]
      summary: Connect to terminal via WebSocket
      operationId: connectTerminalWs
      description: >
        Upgrade to WebSocket. Each message is a JSON frame:
        `{"type":"input","data":"<base64>"}` or
        `{"type":"output","data":"<base64>"}`.
      responses:
        '101':
          description: Switching Protocols (WebSocket upgrade)
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
components:
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
      description: Project identifier
...
    terminalId:
      name: terminalId
      in: path
      required: true
      schema:
        type: string
      description: Terminal identifier
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: resource not found

    # ── Projects ──────────────────────────────────────────────────────────────
    ProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: my-project
        description:
          type: string
        rootPath:
          type: string
          description: Absolute workspace path on the agent host
          example: /workspace/my-project
        labels:
          type: object
          additionalProperties:
            type: string
    ProjectResponse:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        rootPath:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer

    # ── Sessions ──────────────────────────────────────────────────────────────
...
    TerminalCreateRequest:
      type: object
      properties:
        shell:
          type: string
          description: Shell executable path
          example: /bin/bash
        cwd:
          type: string
          description: Initial working directory
          example: /workspace/my-project
        env:
          type: object
          additionalProperties:
            type: string
          description: Extra environment variables
        cols:
          type: integer
          description: Initial terminal width in columns
          default: 80
        rows:
          type: integer
          description: Initial terminal height in rows
          default: 24
    TerminalResponse:
      type: object
      required: [id, status, cols, rows, createdAt]
      properties:
        id:
          type: string
        shell:
          type: string
        cwd:
          type: string
        status:
          type: string
          enum: [running, exited]
        cols:
          type: integer
        rows:
          type: integer
        pid:
          type: integer
        createdAt:
          type: string
          format: date-time
    TerminalListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TerminalResponse'
    TerminalResizeRequest:
      type: object
      required: [cols, rows]
      properties:
        cols:
          type: integer
          minimum: 1
        rows:
          type: integer
          minimum: 1